Made "the Home page"
To give an idea about the page structure, I prepared a basic main page in which there is not any content now. Just a filler text "Welcome to Home Page" and a navbar which only has 5 countries button. created color variables in globals.css to change the colors easily when needed.