fix data validation
I initially tried to create the database with data validation queries, but it was problematic because the existing data didn’t meet some of the validation rules. To resolve this, I modified the approach to use ALTER TABLE to apply the validations. Now, the new database is created by executing the updated validation queries.